chore: Disable tests autorun
