+++ src/SamplesApp/SamplesApp.Skia.Gtk/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "SamplesApp.Skia.Gtk": {
       "commandName": "Project",
-      "commandLineArgs": "--auto-screenshots=C:\\temp\\screenshots",
+      "commandLineArgs": "--_auto-screenshots=C:\\temp\\screenshots",
       "nativeDebugging": false
     }
   }
